requirements.txt
fastapi
pydantic
pydantic_settings
uvicorn
httpx
pytest
tortoise-orm[asyncpg]
aerich
